Refactor data-parser into specialized modules
